Client: move memory stats from execution to service module (to be execution-independent)
